requirements.txt
Flask==3.0.3
googlesearch-python==1.2.4
Werkzeug==3.0.3
